Code cleanup, removed useless functionality. Added more settings to config. API response now provides metadata for uploaded image.

+++ addtodanbooru.py
@@ -3,7 +3,6 @@ import sys
 import time
 import hashlib
 import datetime
-import logwriter
 import imgdatabase
 import updateschedule
 import saucenao
@@ -16,6 +15,8 @@ elif sys.platform == "win32":
     sys.path.append("P:/repos/DanbooruAPI/")
 
 import danbooru
+
+IS_DEBUG = hasattr(sys, 'gettrace') and sys.gettrace() is not None 
 
 danAPI = danbooru.API()
 imgDB = imgdatabase.database()
@@ -41,67 +42,57 @@ blacklisted_terms = [
 
 
 def search_image(full_path:str):
-    results = imgDB.execute_query(f"SELECT image_uid FROM Images WHERE full_path=?", [full_path])
-    return results
+    return imgDB.execute_query(f"SELECT image_uid FROM Images WHERE full_path=?", [full_path])
 
 
-def image_query(full_path) -> int:
+def insert_image(full_path):
     file_name, ext = os.path.splitext(os.path.split(full_path)[1])
-    return ("INSERT INTO Images (full_path, file_name, ext) VALUES (?, ?, ?);", [full_path, file_name, ext])
+    img_qry, img_params = ("INSERT INTO Images (full_path, file_name, ext) VALUES (?, ?, ?);", [full_path, file_name, ext])
+    return imgDB.execute_change(img_qry, img_params)
 
 
-def results_query(image_id, site_flag, illust_id, similarity):
-    return ("INSERT INTO Saucenao_Results (image_uid, site_flag, site_id, similarity, status) VALUES (?, ?, ?, ?, 0);", 
-            [image_id, site_flag, illust_id, similarity])
+def insert_result(image_uid:int, site_flag:int, illust_id:int, similarity:float):
+    rst_qry, rst_params = ("INSERT INTO Saucenao_Results (image_uid, site_flag, site_id, similarity, status) VALUES (?, ?, ?, ?, 0);", 
+                           [image_uid, site_flag, illust_id, similarity])
+    imgDB.execute_change(rst_qry, rst_params)
 
 
-def insert_result(full_path:str, site_flag:int, illust_id:int, similarity:float, found:bool, image_uid = None):
-    if image_uid == None:
-        img_qry, img_params = image_query(full_path)
-        image_uid = imgDB.execute_change(img_qry, img_params)
-
-    if found:
-        rst_qry, rst_params = results_query(image_uid, site_flag, illust_id, similarity)
-        imgDB.execute_change(rst_qry, rst_params)
-    
-    return image_uid
-
-def skip_file(fname:str):
+def skip_file(full_path:str):
     """Skip non-files, files that have already been searched before, not of a valid extension, or has blacklisted terms (mainly for AI art)."""
-    if not os.path.isfile(fname):
-        return True
-    elif any(search_image(fname)):
-        return True
-    elif not os.path.splitext(fname)[1] in saucenao.API.get_allowed_extensions():
-        return True
-    elif any(bl in fname for bl in blacklisted_terms):
-        return True
-    
-    return False
+    return (not os.path.isfile(full_path) or
+           not os.path.splitext(full_path)[1] in saucenao.API.get_allowed_extensions() or 
+           any(bl in full_path for bl in blacklisted_terms) or
+           any(search_image(full_path)))
 
 
-def write_to_log(msg, name):
+def format_time(val):
+    return (f'{val}' if val > 9 else f'0{val}')
+
+
+def write_log(msg, name):
     dt = datetime.datetime.now()
-    strdt = str(dt.date()) + "_" + str(dt.hour) + str(dt.minute)
+    formated_time = f"{dt.date()}_{format_time(dt.hour)}:{format_time(dt.minute)}"
     path = os.path.expanduser("~/Desktop/saucenao_logs/")
     if not os.path.exists(f"{path}"):
         os.makedirs(f"{path}")
-    logwriter.write_log(path + f"{name}_{strdt}.log", msg, "w")
+    with open(f"{path}{name}_{formated_time}.log", "w") as f:
+        f.write(msg)
 
 
 def __add_favorite(full_path:str, file_name:str, illust_id:int, similarity:int = None):
     danAPI.add_favorite(illust_id)
     print(f"Match found ({similarity or 'via md5'}%): {file_name} favorited to {illust_id}, file removed.")
-    os.remove(full_path)
+    if not IS_DEBUG:
+        os.remove(full_path)
+
 
 
 def __check_danbooru(full_path:str, file_name:str) -> bool:
-    """Checks file for source/MD5 match on Danbooru."""
+    """Checks file for MD5 match on Danbooru."""
 
     params = {}
     with open(full_path, "rb") as file:
-        # NOTE: {"md5": f"{img_md5}"} <- DON'T USE. This is a very iffy tag, it only supplies 1 result if found and 
-        # if nothing is found it gives a 404 error. "tags" param is safer as if it finds nothing it returns empty.
+        # {"md5": f"{img_md5}"} <- NOTE: DON'T USE. No results gives 404 error. "tags" is safer, returns empty if nothing found.
         params = {"tags": f"md5:{hashlib.md5(file.read()).hexdigest()}"}
     
     json_data = danAPI.get_posts(params)
@@ -113,44 +104,41 @@ def __check_danbooru(full_path:str, file_name:str) -> bool:
     return False
 
 
-def __process_results(full_path, file_name, response, threshold, db_bitmask:int) -> list[Result]:
-    """summary: Extract file data and send to saucenao REST API. Log low similarity results to {log_name} for later use.
+def __process_results(full_path, file_name, response, high_threshold, low_threshold, db_bitmask, image_data):
+    """summary: Extract file data and send to saucenao REST API. Log low similarity results to database."""
 
-    Args:
-        file_path (str): Full path to locate the file name.
-        fname (str): Just the name of the file.
-        threshold (int): Minimum threshold to detect if an image matches.
-        db_bitmask (int): Flags set that will pull from certain websites in request.
-        log_name (str): Name of the log file.
-    """
-    image_uid = None
-    found = False
-    for r in response["results"]:
-        result = Result(db_bitmask, r)
-        if result.header.similarity >= threshold:
-            found = True
-            # 93 and above are safe bets that the images are exactly the same (sans possible resolution differences)
-            if result.header.similarity > 93:
-                try:
+    # Get a list of all results above the minimum threshold.
+    results:list[Result] = list(filter(lambda r: r.header.similarity > low_threshold, [Result(db_bitmask, r) for r in response["results"]]))
+    image_uid = insert_image(full_path)
+    if not any(results):
+        print(f"No valid results found for {file_name}.")
+        return
+
+    for result in results:
+        if result.header.similarity > high_threshold:
+            try:
+                # Prevent trading down for a lower res image. Give a slight margin of 5%.
+                width, height = image_data["dimensions"]
+                post = danAPI.get_post(result.data.dan_id)
+                if ((width+height) * .95) > (post["image_width"] + post["image_height"]):
+                    print("Danbooru resolution smaller, keeping original image.")
+                else:
                     __add_favorite(full_path, file_name, result.data.dan_id, result.header.similarity)
-                    break
-                except Exception as e:
-                    image_uid = insert_result(full_path, saucenao.API.DBMask.index_danbooru, result.data.dan_id, result.header.similarity, found, image_uid)
-                    raise Exception(f"""Failed to connect to Danbooru's API: {e}\n
-                                    {file_name} added to database but will need to be manually reviewed via 'check-results'.\n
-                                    Check that you have your Danbooru username/API Key as env variables.
-                                    If using a VPN this could be causing 403 errors.""")
-            # Anything lower will need to be double checked via 'check-results'.
-            else:
-                image_uid = insert_result(full_path, saucenao.API.DBMask.index_danbooru, result.data.dan_id, result.header.similarity, found, image_uid)
-                print(f"{file_name} didn't meet similarity quota: {result.header.similarity}%, added record.")
-    if not found:
-        insert_result(full_path, 0, 0, 0, found)
-        print(f"No results found for {file_name}.")
+                    # Remove record as well since we won't need it.
+                    imgDB.execute_change("DELETE FROM Images WHERE image_uid=?",[image_uid])
+                    break # If image is determined a good enough match, move on.
+            except Exception as e:
+                insert_result(image_uid, saucenao.API.DBMask.index_danbooru, result.data.dan_id, result.header.similarity)
+                raise Exception(f"{e}")
+        # Anything lower will need to be double checked via 'check-results'.
+        else:
+            insert_result(image_uid, saucenao.API.DBMask.index_danbooru, result.data.dan_id, result.header.similarity)
+            print(f"{file_name} didn't meet similarity quota: {result.header.similarity}%, added record.")
 
-def add_to_danbooru(directory:str, recursive:bool, threshold:int, schedule:bool, hash_only:bool):
+
+def add_to_danbooru(directory:str, recursive:bool, high_threshold:int, low_threshold:int, schedule:bool):
     db_bitmask = int(saucenao.API.DBMask.index_danbooru)
-    sauceAPI = saucenao.API(db_bitmask, threshold)
+    sauceAPI = saucenao.API(db_bitmask, low_threshold)
 
     # Setup list of files to be searched
     all_files: list[str] = []
@@ -165,22 +153,24 @@ def add_to_danbooru(directory:str, recursive:bool, threshold:int, schedule:bool,
                 continue
 
             file_name = os.path.split(full_path)[1]
-            # Try to save a search by checking Danbooru first. Saucenao's API has a 100 search limit, Dan's doesn't, so doesn't hurt to check.
-            if not __check_danbooru(full_path, file_name) and not hash_only:
+            # Saucenao has a 100 daily search limit, but Dan doesn't. We can save searches by checking image md5 on Dan.
+            if not __check_danbooru(full_path, file_name):
                 api_response = sauceAPI.send_request(full_path)
-                if any(api_response) and int(api_response['header']['results_returned']) > 0:
-                    short_remaining = api_response["header"]["short_remaining"]
-                    long_remaining = api_response["header"]["long_remaining"]
+                response = api_response["response"]
+                image_data = api_response["image"]
+                if any(response) and int(response["header"]["results_returned"]) > 0:
+                    short_remaining = response["header"]["short_remaining"]
+                    long_remaining = response["header"]["long_remaining"]
                     print(f"Remaining Searches 30s|24h: {short_remaining}|{long_remaining}")
 
-                    __process_results(full_path, file_name, api_response, threshold, db_bitmask)
+                    __process_results(full_path, file_name, response, high_threshold, low_threshold, db_bitmask, image_data)
 
                     # Check remaining searches.
                     if long_remaining <= 0:
                         print("Reached daily search limit, unable to process more request at this time.")
                         break
                     elif short_remaining <= 0:
-                        print('Out of searches for this 30 second period. Sleeping for 25 seconds...')
+                        print("Out of searches for this 30 second period. Sleeping for 25 seconds...")
                         time.sleep(25)
                 else:
                     insert_result(full_path, 0, 0, 0)
@@ -189,11 +179,11 @@ def add_to_danbooru(directory:str, recursive:bool, threshold:int, schedule:bool,
         else:
             msg = f"All files scanned for {directory}"
             name = "saucenao_completed"
-            write_to_log(msg, name)
+            write_log(msg, name)
     except Exception as e:
         error_msg = (str(e))
         name = "saucenao_error"
-        write_to_log(error_msg, name)
+        write_log(error_msg, name)
         print(f"{Fore.RED}ERROR: {error_msg}{Style.RESET_ALL}")
 
     # Once finished, set the crontab job to the ending time, this way there will be ample time to refresh the usages.
